Some tweaks, added (not working) glibmm

+++ build.py
@@ -117,6 +117,9 @@ class Project(object):
         if os.path.exists(self.build_dir):
             print_debug("directory %s already exists" % (self.build_dir,))
             self.update_build_dir()
+            if os.path.exists(self.patch_dir):
+                print_log("Copying modified and new files from %s to %s" % (self.patch_dir, self.build_dir))
+                self.builder.copy_all(self.patch_dir, self.build_dir, changed_only=True)
         else:
             self.unpack()
             if os.path.exists(self.patch_dir):
@@ -427,7 +430,7 @@ class Project_glib(Tarball, Project):
             )
 
     def build(self):
-        self.exec_msbuild(r'build\win32\vs%(vs_ver)s\glib.sln /p:PythonPath=%(python_dir)s')
+        self.exec_msbuild(r'build\win32\vs%(vs_ver)s\glib.sln')
         self.install(r'.\COPYING share\doc\glib')
 
 Project.add(Project_glib())
@@ -551,7 +554,7 @@ class Project_libepoxy(Tarball, Project):
         Project.__init__(self,
             'libepoxy',
             archive_url = 'https://github.com/anholt/libepoxy/releases/download/v1.3.1/libepoxy-1.3.1.tar.bz2',
-            patches = '0001-MSVC-Builds-Support-PACKED.patch',
+            patches = ['0001-MSVC-Builds-Support-PACKED.patch'],
             )
 
     def build(self):
@@ -893,6 +896,10 @@ Project.add(MercurialCmakeProject('pygobject', repo_url='git+ssh://git@github.co
 Project.add(MercurialCmakeProject('pygtk', repo_url='git+ssh://git@github.com:muntyan/pygtk-gtk-win32.git', dependencies = ['gtk', 'pycairo', 'pygobject']))
 
 
+Project.add(CmakeProject('libsigc++', archive_url = 'https://download.gnome.org/sources/libsigc++/2.8/libsigc++-2.8.0.tar.xz'))
+Project.add(CmakeProject('glibmm', archive_url = 'http://ftp.gnome.org/pub/GNOME/sources/glibmm/2.46/glibmm-2.46.3.tar.xz', dependencies = ['libsigc++', 'glib']))
+
+
 #========================================================================================================================================================
 
 global_verbose = False
@@ -939,8 +946,8 @@ class Builder(object):
         self.x86 = opts.platform == 'Win32'
         self.x64 = not self.x86
 
-        self.msbuild_opts = '/nologo /p:Platform=%(platform)s %(msbuild_opts)s ' % \
-            dict(platform=opts.platform, configuration=opts.configuration, msbuild_opts=opts.msbuild_opts)
+        self.msbuild_opts = '/nologo /p:Platform=%(platform)s /p:PythonPath=%(python_dir)s %(msbuild_opts)s ' % \
+            dict(platform=opts.platform, configuration=opts.configuration, msbuild_opts=opts.msbuild_opts, python_dir=self.opts.python_dir)
 
         if global_verbose:
             self.msbuild_opts += ' /v:normal'
@@ -1071,20 +1078,29 @@ class Builder(object):
         self.__project = None
 
         print_debug("copying %s to %s" % (proj.pkg_dir, self.gtk_dir))
-        self.copy_all(proj.pkg_dir, self.gtk_dir)
+        self.copy_all(proj.pkg_dir, self.gtk_dir, changed_only=True)
         shutil.rmtree(proj.pkg_dir, ignore_errors=True)
 
     def make_dir(self, path):
         if not os.path.exists(path):
             os.makedirs(path)
 
-    def copy_all(self, srcdir, destdir):
+    def copy_all(self, srcdir, destdir, changed_only=False):
         self.make_dir(destdir)
         for f in glob.glob('%s\\*' % (srcdir,)):
-            self.__copy_to(f, destdir)
+            self.__copy_to(f, destdir, changed_only=changed_only)
 
-    def __copy_to(self, src, destdir):
-        #print_debug("__copy_to %s %s" % (src, destdir))
+    def __is_file_different(self, src, dst):
+        try:
+            s1 = os.stat(src)
+            s2 = os.stat(dst)
+            #print src, s1
+            #print dst, s2
+            return s1.st_size != s2.st_size or s1.st_mtime != s2.st_mtime
+        except:
+            return True
+
+    def __copy_to(self, src, destdir, changed_only=False):
         self.make_dir(destdir)
         for f in glob.glob(src):
             if os.path.isdir(f):
@@ -1092,9 +1108,11 @@ class Builder(object):
                 dest_subdir = os.path.join(destdir, name)
                 self.make_dir(dest_subdir)
                 for item in os.listdir(f):
-                    self.__copy_to(os.path.join(f, item), dest_subdir)
+                    self.__copy_to(os.path.join(f, item), dest_subdir, changed_only=changed_only)
             else:
-                shutil.copy2(f, destdir)
+                if not changed_only or self.__is_file_different(f, os.path.join(destdir, os.path.basename(f))):
+                    print_debug("__copy_to: copying %s %s" % (f, destdir))
+                    shutil.copy2(f, destdir)
 
     def __copy(self, src, destdir):
         if os.path.isdir(src):
@@ -1146,13 +1164,13 @@ class Builder(object):
         for f in args[:-1]:
             src = os.path.join(self.__sub_vars(build_dir), self.__sub_vars(f))
             print_debug("copying %s to %s" % (src, dest))
-            self.__copy_to(src, dest)
+            self.__copy_to(src, dest, changed_only=True)
 
     def install_dir(self, build_dir, pkg_dir, src, dest):
         src = os.path.join(build_dir, self.__sub_vars(src))
         dest = os.path.join(pkg_dir, self.__sub_vars(dest))
         print_debug("copying %s content to %s" % (src, dest))
-        self.copy_all(src, dest)
+        self.copy_all(src, dest, changed_only=True)
 
     def exec_msys(self, args, working_dir=None):
         self.__execute(args, working_dir=working_dir, add_path=os.path.join(self.opts.msys_dir, 'usr', 'bin'))
